Separate compareFeatures() into new file
