+++ test/fix-feature-order.js
@@ -19,17 +19,7 @@
 const fs = require('fs');
 const path = require('path');
 
-const compareFeatures = (a,b) => {
-  if (a == '__compat') return -1;
-  if (b == '__compat') return 1;
-  
-  const wordA = /^[a-zA-Z](\w|-)*$/.test(a);
-  const wordB = /^[a-zA-Z](\w|-)*$/.test(b);
-
-  if (wordA && wordB) return a.localeCompare(b, 'en');
-  if (wordA || wordB) return (wordA && -1) || 1;
-  return a.localeCompare(b, 'en');
-}
+const compareFeatures = require('./compare-features.js');
 
 function orderFeatures(key, value) {
   if (value instanceof Object && '__compat' in value) {
